pyproject.toml
[tool.poetry]
name = "morph-starter-apps"
version = "0.1.0"
description = ""
authors = ["shibatanaoto <naoto.shibata510@gmail.com>"]
readme = "README.md"
packages = [
    { include = "src" }
]

[tool.poetry.dependencies]
python = "<3.13,>=3.9"
morph-data = "0.1.2"
langchain = "0.3.16"
langchain-core = "0.3.32"
langchain-openai = "0.3.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
